Incorporate interpreter.js into Region class
